Renamed hmac.py to hmac_utils.py to avoid module name conflict
